refactor: extract duplicated SortableTableCell into objectUtils
Moved the SortableTableCell component definition from Kinds.jsx and
Namespaces.jsx into the existing components/utils/objectUtils.js module.
Both pages now import it from there, removing the duplication.
